requirements/local.txt
-r base.txt

Werkzeug==1.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.4  # https://github.com/gotcha/ipdb
#psycopg2-binary==2.8.6  # https://github.com/psycopg/psycopg2
mysqlclient
sqlparse
# Testing
# ------------------------------------------------------------------------------
mypy==0.790  # https://github.com/python/mypy
django-stubs==1.7.0  # https://github.com/typeddjango/django-stubs
pytest==6.1.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==3.3.1  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2020.9.1 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.4  # https://github.com/PyCQA/flake8
flake8-isort==4.0.0  # https://github.com/gforcada/flake8-isort
coverage==5.3  # https://github.com/nedbat/coveragepy
black==20.8b1  # https://github.com/ambv/black
pylint-django==2.3.0  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==2.9.3  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.1.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.2  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.1.0  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.1.0  # https://github.com/pytest-dev/pytest-django
django-test-plus
